Move numlen(1) to sway/util.c
