+++ sway/util.c
@@ -3,3 +3,13 @@
 int wrap(int i, int max) {
 	return ((i % max) + max) % max;
 }
+
+int numlen(int n) {
+	if (n >= 1000000) return 7;
+	if (n >= 100000) return 6;
+	if (n >= 10000) return 5;
+	if (n >= 1000) return 4;
+	if (n >= 100) return 3;
+	if (n >= 10) return 2;
+	return 1;
+}
